+++ src/components/common/modal/ButtonsModal/style.ts
